Unnecessarily cryptic commit message: turns out there is no such thing as a free lunch.
Less cryptic: I have hit a dead end in trying to predict future movement of stock prices.
The neural network approach refuses to work, regardless of which study I attempt to emulate.
The SVM approach is similarly unproductive.

At the next meeting I will propose mothballing this project and moving on to either a probabilistic modeling approach or a portfolio optimization tool